Added the BE networking module
Supersedes https://github.com/Legacy-Fabric/fabric/pull/35

 Changes to be committed:
	new file:   fabric-networking-blockentity-v0/build.gradle
	renamed:    unfinished/fabric-networking-blockentity-v0/src/main/java/net/fabricmc/fabric/api/block/entity/BlockEntityClientSerializable.java -> fabric-networking-blockentity-v0/src/main/java/net/fabricmc/fabric/api/block/entity/BlockEntityClientSerializable.java
	renamed:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/mixin/object/builder/MaterialBuilderHooks.java -> fabric-networking-blockentity-v0/src/main/java/net/fabricmc/fabric/mixin/networking/blockentity/BlockEntityAccessor.java
	renamed:    unfinished/fabric-networking-blockentity-v0/src/main/java/net/fabricmc/fabric/mixin/networking/blockentity/MixinBlockEntity.java -> fabric-networking-blockentity-v0/src/main/java/net/fabricmc/fabric/mixin/networking/blockentity/MixinBlockEntity.java
	renamed:    unfinished/fabric-networking-blockentity-v0/src/main/java/net/fabricmc/fabric/mixin/networking/blockentity/MixinClientPlayNetworkHandler.java -> fabric-networking-blockentity-v0/src/main/java/net/fabricmc/fabric/mixin/networking/blockentity/MixinClientPlayNetworkHandler.java
	renamed:    unfinished/fabric-networking-blockentity-v0/src/main/resources/assets/fabric-networking-blockentity-v0/icon.png -> fabric-networking-blockentity-v0/src/main/resources/assets/fabric-networking-blockentity-v0/icon.png
	renamed:    unfinished/fabric-networking-blockentity-v0/src/main/resources/fabric-networking-blockentity-v0.mixins.json -> fabric-networking-blockentity-v0/src/main/resources/fabric-networking-blockentity-v0.mixins.json
	renamed:    unfinished/fabric-networking-blockentity-v0/src/main/resources/fabric.mod.json -> fabric-networking-blockentity-v0/src/main/resources/fabric.mod.json
	modified:   settings.gradle
	deleted:    unfinished/fabric-networking-blockentity-v0/build.gradle
	deleted:    unfinished/fabric-object-builders-v0/build.gradle
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/api/block/BlockSettingsExtensions.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/api/block/FabricBlockSettings.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/api/block/FabricMaterialBuilder.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/api/entity/EntityTrackingRegistry.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/api/entity/FabricEntityTypeBuilder.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/api/event/registry/BlockConstructedCallback.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/api/event/registry/ItemConstructedCallback.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/impl/object/builder/FabricEntityType.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/mixin/object/builder/BlockSettingsHooks.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/mixin/object/builder/MixinBlock.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/mixin/object/builder/MixinItem.java
	deleted:    unfinished/fabric-object-builders-v0/src/main/resources/assets/fabric-object-builders-v0/icon.png
	deleted:    unfinished/fabric-object-builders-v0/src/main/resources/fabric-object-builders-v0.mixins.json
	deleted:    unfinished/fabric-object-builders-v0/src/main/resources/fabric.mod.json
