+++ fabric-networking-blockentity-v0/build.gradle
@@ -1,0 +1,7 @@
+archivesBaseName = "fabric-networking-blockentity-v0"
+version = getSubprojectVersion(project, "0.1.0")
+
+dependencies {
+	implementation project(path: ':fabric-api-base', configuration: 'dev')
+}
+
