+++ fabric-networking-blockentity-v0/src/main/resources/assets/fabric-networking-blockentity-v0/icon.png
