+++ fabric-networking-blockentity-v0/src/main/resources/fabric-networking-blockentity-v0.mixins.json
@@ -1,0 +1,15 @@
+{
+  "required": true,
+  "package": "net.fabricmc.fabric.mixin.networking.blockentity",
+  "compatibilityLevel": "JAVA_8",
+  "mixins": [
+    "BlockEntityAccessor",
+    "MixinBlockEntity"
+  ],
+  "client": [
+    "MixinClientPlayNetworkHandler"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
