+++ settings.gradle
@@ -13,18 +13,19 @@ pluginManagement {
 rootProject.name = "fabric-api"
 
 include 'fabric-api-base'
+include 'fabric-armor-api-v1'
+include 'fabric-commands-v0'
 include 'fabric-content-registries-v1'
 include 'fabric-crash-report-info-v1'
-include 'fabric-resource-loader-v0'
 include 'fabric-events-interaction-v0'
 include 'fabric-events-lifecycle-v0'
-include 'fabric-networking-v0'
+include 'fabric-gamerule-api-v1'
 include 'fabric-item-groups-v0'
 include 'fabric-key-binding-api-v1'
-include 'fabric-commands-v0'
-include 'fabric-rendering-v1'
-include 'fabric-armor-api-v1'
-include 'fabric-gamerule-api-v1'
 include 'fabric-lifecycle-events-v1'
-include 'fabric-recipes-v1'
+include 'fabric-networking-blockentity-v0'
+include 'fabric-networking-v0'
 include 'fabric-object-builder-api-v1'
+include 'fabric-recipes-v1'
+include 'fabric-rendering-v1'
+include 'fabric-resource-loader-v0'
