--- unfinished/fabric-networking-blockentity-v0/build.gradle
@@ -1,7 +1,0 @@
-archivesBaseName = "fabric-networking-blockentity-v0"
-version = getSubprojectVersion(project, "0.2.3")
-
-dependencies {
-	compile project(path: ':fabric-api-base', configuration: 'dev')
-}
-
