--- unfinished/fabric-networking-blockentity-v0/src/main/resources/fabric.mod.json
@@ -1,26 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-networking-blockentity-v0",
-  "name": "Fabric Networking Block Entity (v0)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-networking-blockentity-v0/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.4.0",
-    "fabric-api-base": "*"
-  },
-  "description": "Networking hooks for block entities.",
-  "mixins": [
-    "fabric-networking-blockentity-v0.mixins.json"
-  ]
-}
