--- unfinished/fabric-object-builders-v0/build.gradle
@@ -1,7 +1,0 @@
-archivesBaseName = "fabric-object-builders"
-version = getSubprojectVersion(project, "0.1.4")
-
-dependencies {
-    compile project(path: ':fabric-api-base', configuration: 'dev')
-    compile project(path: ':fabric-mining-levels-v0', configuration: 'dev')
-}
