--- unfinished/fabric-object-builders-v0/src/main/resources/fabric-object-builders-v0.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.object.builder",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [
-    "BlockSettingsHooks",
-    "MaterialBuilderHooks",
-    "MixinBlock",
-    "MixinItem"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
