--- unfinished/fabric-object-builders-v0/src/main/resources/fabric.mod.json
@@ -1,27 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-object-builders-v0",
-  "name": "Fabric Object Builders (v0)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-object-builders-v0/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.4.0",
-    "fabric-api-base": "*",
-    "fabric-mining-levels-v0": "*"
-  },
-  "description": "Builders for objects vanilla has locked down.",
-  "mixins": [
-    "fabric-object-builders-v0.mixins.json"
-  ]
-}
